Fixed never ending decodeLengthPrefixC
Bug discovered with property testing but was mostly harmless in
a with a real socket Producer since the conduit was torn down anyway when
the exception from the closed socket bubbled up and killed the thread which
handled the connection.

We now don't have an annoying error message thrown to stderr whenever a
connection was closed by the peer.